moved all the main functions into the global state that way i can use it inside both Mentor & Mentee components
