+++ pages/abis/Decentee.json
@@ -16573,12 +16573,12 @@
     "5777": {
       "events": {},
       "links": {},
-      "address": "0x340437B0bE47D94485217E6a0d44C15C7C46C9ac",
-      "transactionHash": "0x593f8983cc8bfcce835f6218928ab7779f188f7cc0c9e1d05cec0c1a5a0c36eb"
+      "address": "0xA2549D128a34E3b7E10cF4d496A8aFa4761c13a5",
+      "transactionHash": "0xa0e459aeda62d260894d52cbcb8b9e306296427f32e1b1415674f6ea8fa69055"
     }
   },
   "schemaVersion": "3.4.1",
-  "updatedAt": "2021-06-28T19:14:41.180Z",
+  "updatedAt": "2021-07-14T02:37:28.409Z",
   "networkType": "ethereum",
   "devdoc": {
     "kind": "dev",
