+++ pages/abis/Migrations.json
@@ -2564,12 +2564,12 @@
     "5777": {
       "events": {},
       "links": {},
-      "address": "0xA17b7C30332E88aF6A17419517Fe5a5a117b05bE",
-      "transactionHash": "0xe4a874dad896219ab76131ebbac881cca752ea1fa3b5e3661711814173a5a00e"
+      "address": "0xEcd7988F5b352b80B25A02BbD92408279bd94C36",
+      "transactionHash": "0xfe6d04be5d18ed2d24c1dbfb42f6e514e67fca0a2de3191018d14eb78af7649e"
     }
   },
   "schemaVersion": "3.4.1",
-  "updatedAt": "2021-06-28T19:14:41.229Z",
+  "updatedAt": "2021-07-14T02:37:28.419Z",
   "networkType": "ethereum",
   "devdoc": {
     "kind": "dev",
